+++ src/deluge/model/clip/instrument_clip.h
@@ -107,6 +107,9 @@ public:
 
 	ArpeggiatorSettings arpSettings; // Not valid for Kits
 	int32_t arpeggiatorRate;
+	int32_t arpeggiatorRatchetProbability;
+	int32_t arpeggiatorRatchetAmount;
+	int32_t arpeggiatorSequenceLength;
 	int32_t arpeggiatorGate;
 
 	ParamManagerForTimeline backedUpParamManagerMIDI;